feat: 🎸 added dynamic cover img for the frame
