+++ src/app/frames/[netname]/[id]/app.tsx
@@ -1,6 +1,8 @@
 'use client';
 
+import sdk from '@farcaster/frame-sdk';
 import dynamic from 'next/dynamic';
+import { useEffect, useState } from 'react';
 
 const Claims = dynamic(() => import('@/components/frame/claims/claims'), {
   ssr: false,
@@ -13,5 +15,15 @@ export default function App({
   bountyId: string;
   chainId: string;
 }) {
+  const [isSDKLoaded, setIsSDKLoaded] = useState(false);
+  useEffect(() => {
+    const load = async () => {
+      sdk.actions.ready();
+    };
+    if (sdk && !isSDKLoaded) {
+      setIsSDKLoaded(true);
+      void load();
+    }
+  }, [isSDKLoaded]);
   return <Claims chainId={chainId} bountyId={bountyId} />;
 }
